Refactor stars url creation
