+++ src/markdown.rs
@@ -14,6 +14,7 @@ pub async fn render(data: Data) -> Result<()> {
 
     let mut tera = Tera::new("templates/*.tmpl")?;
     tera.register_filter("to_snake_case", utils::to_snake_case_filter);
+    tera.register_filter("stars_url", utils::stars_url_filter);
 
     render_readme(&tera, &data).await?;
     render_projects(&github, &tera, &data).await?;
